refector code for network and users
